pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "beautifulsoup4>=4.13.3",
    "discord-py>=2.5.2",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.1",
    "sqlalchemy>=2.0.38",
    "trafilatura>=2.0.0",
    "twilio>=9.4.6",
]
